Remove PySpectrometer2-Picam2 and filelist.txt
The file filelist.txt is useless in general, so we can safely remove it.
The PySpectrometer2-Picam2 script is for controling a sensor directly
attached to the camera sensor interface of the Raspberry Pi.
The little garden spectrometer is always a USB-connected device,
so we don't need the Picam2 code.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>